+++ Engine/Shared/Interfaces/IEntityCommandBuffer.cs
@@ -7,7 +7,7 @@ namespace Shared.Interfaces;
     /// </summary>
     public interface IEntityCommandBuffer
     {
-        void CreateObject(ObjectType objectType, int x = 0, int y = 0, int z = 0);
+        void CreateObject(ObjectType objectType, long x = 0, long y = 0, long z = 0);
         void DestroyObject(IGameObject obj);
         void AddComponent<T>(IGameObject obj, T component) where T : class, IComponent;
         void RemoveComponent<T>(IGameObject obj) where T : class, IComponent;
